# rag/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "literag",
    "version": "0.1.0"
  },
  "paths": {
    "/documents/upload": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Upload Document",
        "operationId": "upload_document_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List Documents",
        "operationId": "list_documents_documents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentSummary"
                  },
                  "type": "array",
                  "title": "Response List Documents Documents Get"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Document",
        "operationId": "get_document_documents__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}/chunks": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List Chunks",
        "operationId": "list_chunks_documents__document_id__chunks_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Chunk"
                  },
                  "title": "Response List Chunks Documents  Document Id  Chunks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}/chunks/{chunk_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Chunk",
        "operationId": "get_chunk_documents__document_id__chunks__chunk_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "chunk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chunk Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chunk"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}/relations": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List Chunk Relations",
        "operationId": "list_chunk_relations_documents__document_id__relations_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Relation"
                  },
                  "title": "Response List Chunk Relations Documents  Document Id  Relations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Rebuild Relations",
        "operationId": "rebuild_relations_documents__document_id__relations_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelationsBuildResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}/entities": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List Entities",
        "operationId": "list_entities_documents__document_id__entities_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Entity"
                  },
                  "title": "Response List Entities Documents  Document Id  Entities Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}/entity-relations": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List Entity Relations",
        "operationId": "list_entity_relations_documents__document_id__entity_relations_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityRelation"
                  },
                  "title": "Response List Entity Relations Documents  Document Id  Entity Relations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/query": {
      "post": {
        "tags": [
          "query"
        ],
        "summary": "Query",
        "operationId": "query_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/healthz": {
      "get": {
        "summary": "Health",
        "operationId": "health_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Healthz Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_document_documents_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_documents_upload_post"
      },
      "Chunk": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "chunk_index": {
            "type": "integer",
            "title": "Chunk Index",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "document_id",
          "source",
          "text"
        ],
        "title": "Chunk"
      },
      "DocumentDetail": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "indexed"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "source",
          "chunk_count"
        ],
        "title": "DocumentDetail"
      },
      "DocumentSummary": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "indexed"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "source",
          "chunk_count"
        ],
        "title": "DocumentSummary"
      },
      "Entity": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "id",
          "type"
        ],
        "title": "Entity"
      },
      "EntityRelation": {
        "properties": {
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "relation_type": {
            "type": "string",
            "title": "Relation Type"
          },
          "target_id": {
            "type": "string",
            "title": "Target Id"
          },
          "target_type": {
            "type": "string",
            "title": "Target Type"
          }
        },
        "type": "object",
        "required": [
          "source_id",
          "source_type",
          "relation_type",
          "target_id",
          "target_type"
        ],
        "title": "EntityRelation"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "QueryRequest": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "top_k": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "QueryRequest"
      },
      "QueryResponse": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/QuerySource"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "question",
          "sources"
        ],
        "title": "QueryResponse"
      },
      "QuerySource": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "document_id",
          "source",
          "score",
          "text"
        ],
        "title": "QuerySource"
      },
      "Relation": {
        "properties": {
          "source_chunk_id": {
            "type": "string",
            "title": "Source Chunk Id"
          },
          "target_chunk_id": {
            "type": "string",
            "title": "Target Chunk Id"
          },
          "relation_type": {
            "type": "string",
            "title": "Relation Type"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "source_chunk_id",
          "target_chunk_id",
          "relation_type"
        ],
        "title": "Relation"
      },
      "RelationsBuildResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "task_id",
          "status"
        ],
        "title": "RelationsBuildResponse"
      },
      "UploadResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "filename",
          "chunk_count"
        ],
        "title": "UploadResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}